Add files via upload
The rate increasing population in our country has increased rapidly and also we have an increase in garbage which have increased environmental issue. The dustbin is a container that collects garbage’s or stores items
which recyclable or non-recyclable, decompose and non-decompose. They are usually used in homes, fice, etc, but in case they are full no one is there to clean it and the garbage is spilled out. The surrounding of a dustbin is also conducive for increasing the pollution level. Air pollution due to a dustbin can produce bacteria and viruses which can produce life harmful diseases for humans. Therefore, we have designed a smart dustbin using ARDUINO UNO, an ultrasonic sensor that will sense the item to be thrown in the dustbin and open the lid with the help of the motor. It is an IOT based project that will bring a new and smart way of cleanliness. It is a decent gadget to make your home clean, due to practically all offspring of home consistently make it grimy and spread litter to a great extentby electronics, rappers, and various other things. Since the smart dustbin is additionally intriguing and children make fun with it so it will help to maintain cleanliness in home. It will be applied to various types of waste. Dustbin will open its lid when someone/object is near at some range hen it will wait for a given time period then it will close automatically. Here lid will close when you don’t want to use and it will only open hence required. SMART DUSTBIN USING ARDUINO is an IOT based project. Here we are using Arduino for code execution, for sensing we used an ultrasonic sensor that will open the lid and wait for a few moments. It will bring drastic changes intern of cleanliness with the help of technology. Everything is getting with
smart technology for the betterment of human beings. So this helps in maintaining the environment clean with the help of technology. It is a sensor-based dustbin so it would be easy to access/use for any age group. Our aim is also to make it cost-effective so that many numbers of people can get the benefit from this. And it should be usable to anyone and helpful for them.